feat: move expr_parser and remove module.rs to definitions
