+++ crates/aiken-lang/src/parser/definitions/mod.rs
@@ -1,3 +1,5 @@
+use chumsky::prelude::*;
+
 pub mod constant;
 mod data_type;
 mod function;
@@ -13,3 +15,20 @@ pub use import::parser as import;
 pub use test::parser as test;
 pub use type_alias::parser as type_alias;
 pub use validator::parser as validator;
+
+use super::{error::ParseError, token::Token};
+use crate::ast;
+
+pub fn parser() -> impl Parser<Token, Vec<ast::UntypedDefinition>, Error = ParseError> {
+    choice((
+        import(),
+        data_type(),
+        type_alias(),
+        validator(),
+        function(),
+        test(),
+        constant(),
+    ))
+    .repeated()
+    .then_ignore(end())
+}
